move data tools specific test utlis to separate file
