+++ test/gurka/test_closure_penalty.cc
@@ -1,5 +1,6 @@
 #include "gurka.h"
 #include "test.h"
+#include "datatools/test.h"
 
 #include "baldr/graphreader.h"
 #include "baldr/traffictile.h"
@@ -122,7 +123,7 @@ protected:
                                       [](baldr::GraphReader& reader, baldr::TrafficTile& tile,
                                          uint32_t index, baldr::TrafficSpeed* current) -> void {
                                         (void)reader, (void)tile, (void)index;
-                                        SetLiveSpeed(current, UNKNOWN_TRAFFIC_SPEED_RAW);
+                                        SetLiveSpeed(current, baldr::UNKNOWN_TRAFFIC_SPEED_RAW);
                                       });
   }
