+++ test/gurka/test_costing_with_traffic.cc
@@ -1,5 +1,6 @@
 #include "gurka.h"
 #include "test.h"
+#include "datatools/test.h"
 #include <gtest/gtest.h>
 
 #if !defined(VALHALLA_SOURCE_DIR)
