+++ test/gurka/test_reach.cc
@@ -2,9 +2,8 @@
 #include "loki/reach.h"
 #include "sif/costfactory.h"
 #include "sif/dynamiccost.h"
-#include "test.h"
+#include "datatools/test.h"
 
-#include <boost/format.hpp>
 #include <gtest/gtest.h>
 #include <valhalla/proto/options.pb.h>
 
@@ -125,7 +124,7 @@ TEST_F(TestReach, ReachWithNoClosures) {
   sif::CostFactory factory;
   CostingOptions co;
   co.set_costing(valhalla::auto_);
-  co.set_flow_mask(kDefaultFlowMask);
+  co.set_flow_mask(baldr::kDefaultFlowMask);
   co.set_filter_closures(false);
 
   auto costing = factory.Create(co);
@@ -160,7 +159,7 @@ TEST_F(TestReach, ReachWithClosures) {
 
   CostingOptions co;
   co.set_costing(valhalla::auto_);
-  co.set_flow_mask(kDefaultFlowMask);
+  co.set_flow_mask(baldr::kDefaultFlowMask);
   co.set_filter_closures(true);
   // Check reach for AB - enclosed by closures, so reach should be 0
   {
@@ -221,7 +220,7 @@ TEST_F(TestReach, DISABLED_ReachWithClosures2) {
 
   CostingOptions co;
   co.set_costing(valhalla::auto_);
-  co.set_flow_mask(kDefaultFlowMask);
+  co.set_flow_mask(baldr::kDefaultFlowMask);
   co.set_filter_closures(false);
   auto costing = factory.Create(co);
   auto edge = gurka::findEdgeByNodes(*reader, closure_map.nodes, "K", "L");
