+++ test/gurka/test_recost.cc
@@ -2,6 +2,7 @@
 #include "mjolnir/graphtilebuilder.h"
 #include "sif/recost.h"
 #include "test.h"
+#include "datatools/test.h"
 
 using namespace valhalla;
