+++ test/gurka/test_route_incidents.cc
@@ -1,5 +1,6 @@
 #include "gurka.h"
 #include "test.h"
+#include "datatools/test.h"
 
 using namespace valhalla;
