+++ test/gurka/test_search_filter.cc
@@ -1,5 +1,6 @@
 #include "gurka.h"
 #include "test.h"
+#include "datatools/test.h"
 
 #include <boost/format.hpp>
 #include <gtest/gtest.h>
