+++ test/gurka/test_use_living_streets.cc
@@ -1,5 +1,6 @@
 #include "gurka.h"
 #include "test.h"
+#include "datatools/test.h"
 #include <gtest/gtest.h>
 
 using namespace valhalla;
@@ -46,7 +47,7 @@ protected:
     test::customize_historical_traffic(use_living_streets_map.config, [](baldr::DirectedEdge& e) {
       e.set_free_flow_speed(40);
       e.set_constrained_flow_speed(40);
-      return std::array<float, kBucketsPerWeek>{};
+      return std::array<float, baldr::kBucketsPerWeek>{};
     });
   }
 };
