+++ test/gurka/test_use_tracks.cc
@@ -1,5 +1,6 @@
 #include "gurka.h"
 #include "test.h"
+#include "datatools/test.h"
 #include <gtest/gtest.h>
 
 using namespace valhalla;
